Disable "Add Call" and "Mute" while in emergency callback mode
This is Sprint ECM requirement ANDR-118.

Note that this change affects ECM on all CDMA devices regardless of
carrier.  (This particular requirement is from Sprint, but I don't think
it's *contrary* to any other carrier's requirements, so let's just do it
for any device that uses ECM.)

TESTED: On crespo-s, by temporarily hacking the isPhoneInEcm() method to
return true (so I wouldn't need to actually call 911.)  Also removed the
hack and confirmed that "Add Call" and "Mute" still work normally.

Bug: 4147101
Change-Id: Idfeb79892f6cf6b9878735ad12bb97d52c22c0c2